remove CFLAGS from ccm.skel
CFLAGS in pacman-5.2.2-3 are defined over multiple lines with backslashes
which makes the simple sed call here insufficient to pass custom ones to
the buildroot.

If users want to use a custom CFLAGS, they will need to define them in the
custom makepkg.conf rather than in the config file.