ccm: support two build roots
Define a secondary build root for users wanting to have the option to run a
primary and secondary buildroot. If the primary is busy, ccm will automatically
use the secondary buildroot.  One use case is having the ability to build a
package in the secondary buildroot while the primary buildroot is busy building an
independent package.  Note that this can result in race conditions so use this with caution.